Use MINIMAL_RUNTIME in benchmarks (#10861)
This has been ready for months now but we just forgot to
flip the switch. Almost all tests work with the new runtime,
except for a few that embed files and such. But all the
small ones (where JS size matters) do work and they
benefit by 20%+ shrinking in many cases.

The matrix multiply test had some browser components
like main loop support, that I don't think we need (maybe
in the past we wanted to run this benchmark in the
browser too? but using the main loop would mess up
the timings...) so I removed those too.