JS_MATH option: use Math.* instead of compiled musl (#10821)
This option uses the JS VM's Math.* code instead of compiled
musl for things like cosf etc. This is smaller (on a small testcase
it can save quite a lot, like 25% of total size), but is also much
slower (it calls out of wasm to JS) so it likely isn't that useful except
for cases where speed just doesn't matter. However it may be
nice in wasm2js builds, where there is no cost to calling out of
wasm. (There are still downsides, though - the JS Math.* have
slightly different semantics than libc.)

This simply implements the libc functions as EM_JS calls to JS,
and links that in before libc so it overrides it.

Inspired by AssemblyScript's --use Math=JSMath option.