Benchmark fixes and improvements (#10870)
We had an annoying little bug where we set CFLAGS so that
feature flags work, but some build systems don't append that -
they just replace all their defaults with it. As a result, on the zlib
test we were measuring unoptimized code (lol) which was
over 30% larger and almost half the speed it should be...

Add an LTO mode for emscripten. LTO sometimes helps by a
small amount, but it's mixed, and can also hurt by a little bit too
on code size, due to much more inlining. But in some cases it
helps a lot, like 25% smaller code in fasta, 8% smaller in coremark,
and a 25% speedup on skinning.

Run v8 without liftoff. The baseline compiler can add a lot of
noise since how much time we spend there ends up mattering
a lot (the wasm runs at half speed until the full compiler is done,
and that's noisy).

Also add nicer code for error handling in the zlib benchmark.