Wasm/JS BigInt support (#10860)
This adds a flag WASM_BIGINT which enables this feature, with
which we let the JS VM use a JS BigInt for a wasm i64. In that case
we don't need to legalize i64s into pairs of i32s.

This is fairly straightforward, but we do need to modify the JS code
of each library method that receives or returns an i64.

Tests verify that we can send and receive i64s from wasm to JS,
and also that a dynCall works. This depends on
WebAssembly/binaryen#2726 for that.