Parameterize bullet test
It was running two different and long running builds so it makes
sense to split.  Also rename it since the original reason for adding
the `the` was to put it near the end of the list of test alphabetically
but that no longer makes sense now that we have a parallel test runner.