Remove EMCC_LEAVE_INPUTS_RAW environment variable (#10974)
This was a strange and cross cutting option that is implemented as
an environment variable and only uses by a couple of fastcomp only tests.

It turns out that what these tests were really trying to achieve was
basically `-nostdlib`, so replace the two uses with that option.

There were a few `.ll` tests that are part of `test_zzz_cases` that
required a minor modification to continue to work.  It seems that
for some reason the `linkonce_odr` linkage type on `main` was causing
the globaldce pass to eliminate it, even when `main` was included
in the `-internalize-public-api` list.  I didn't try to get the bottom
of this but instead removed the linkage specifier from the main
functions.