AJSDCE improvement: ignore trivial constructors (#10985)
Closure compiler may leave things like

  new Int8Array(buffer);

since that has a side effect, in this case allocation, and also the
user may have monkey-patched Int8Array to something else.

For us, we depend on those things being not monkey-patched,
and we don't care about an allocation that is never seen. So we can
ignore those.

This reduces code size as you can see in the test updates. Mostly
just noticeable in the very smallest examples.