Link system libraries by default
After this change system library no longer linked on demand based
`.symbols` files, to simply included by default in all cases.

The `.symbols` file logic were originally added as an optimization
would allow only the libraries needed to be considered at link time.

However, this optimization I believe was put in place for fastcomp where
the build time cost (in build) of including a library was significant.

This optimization does not apply with the wasm backend where linking is
fast and `.a` libraries are handled efficiently.  Also are basically no
real programs that don't use libc and for those programs we already have
the `-nostdlib` option.

The `.symbols` files are still used for some optional libraries such
as libhtml, libgl and libal.  However the plan is to also remove that
logic in STRICT mode too.   Once we do that we can avoid having to
run llvm-nm on each input object file.

See #8912.