Refactor test_embind and mark as slow
This test takes an absurd amount of time to compile with fastcomp.

Split out the simple test case that was checking for linker failure.
No need to compile this huge file just to test for that.  This was
also the only case with an expected failures which massively simplies
the code.

Finally move IN_TEST_HARNESS=1 argument down into the loop with all
ther common arguments.

This test takes 30 minutes to run all the cases on fastcomp!  And
around 3 minutues with the llvm backend.