Parameterize test_cmake test. NFC. (#10994)
Remove the `detect` and `prebuild` keys for each generator as they
seem pretty redundnat.  `nmake_detect_error` was just check for the
the build command in the path, which is the default behaviour, and
`prebuild` was just checking for the existance of the Makefile after
cmake runs.

Also add `project` to tests/cmake/emscripten_version/CMakeLists.txt
to avoid warning.