Add 2GB-4GB heap testing (#10811)
It's not easy to test this because (1) we need super-new V8, and
(2) this will allocate over 2GB during testing, which strains the bots'
memory limits and risks OOMs. But we should try to have some
testing at least.

This adds new tests in browser. I think in the long term we want
to test this there anyhow, to make sure this works in actual
browsers and not just a shell environment. For now, we still need
to wait for the V8 update to reach chrome beta, so this doesn't
actually run in a browser, but I still put the test in
browser as that suite runs sequentially, which means
no other test will compete for memory with it, reducing OOM
risks. As we expect to convert this to a browser test eventually
this seems not that bad for now (however, if we want non-browser
2GB+ testing, we should probably add a sequential test mode).

The main new test, test_4GB, allocates over 3GB and also
verifies that we can read and write to that memory in both C
and JS.

This also found a minor bug: without 2GB+ heaps we did not
properly error on trying to allocate 2GB, the value actually
overflowed and we entered the wrong code path. The new
test_2GB_fail verifies that is fixed (the fix is by always
>>> 0 the memory size, which also removes an ifdef,
nicely).