--extern-pre-js / --extern-post-js (#11088)
The new options just prepend/append JS code, after all other
work has been done. That is, these files are not optimized with
everything else, not MODULARIZEd if we do that, not instrumented
with SAFE_HEAP if we do that, etc. They are just prepended/appended
after emcc finishes everything and right before it exits.

A build system could just prepend/append after emcc runs,
but sometimes that's less convenient. For example,
in Binaryen we need something like this now that
MODULARIZE_INSTANCE is gone. So hopefully this helps those
users.