Remove SWAPPABLE_ASM_MODULE setting
For the wasm backend we stopped supporting this correctly back in #9922.
This change means that the top level bindings and the `Module` bindings
for were bound on first use.  This works fro wasm async compilation
but not for replacign running modules.

The only other use was emterpretify which is due to be removed
soon in any case: #11020

We may want to bring back full support for this later, but for now
we don't support it correctly so better to remove it.