When no output suffix is specified, assume executable output (#11108)
We were previously assembly object file output for simple cases
like `emcc -c foo.c -o foo`.   Previously this would attempt to build
a relocatable object file, now it build a JavaScript executable.

This caused the behavior of `emcc -o -` to change. Previously
we silently accepted this because we could create an object
file called `-`.  However, with this change we try to link `-.wasm`
and passing a filename to binaryen that starts with `-` is not
possible, because it thinks its a option.  This seems like reasonable
behaviour the part of binaryen.  Indeed neither gcc or clang support
compiling files that start with `-`.

```
$ clang -c -- -foo.c
error: unknown argument: '-foo.c'
$ gcc -c  -foo.c
gcc: error: unrecognized command line option ‘-foo.c’
```