Default to storing emscripten cache locally
The cache is something that can't really be shared between emscripten
versions.  In fact having a single location for the cache means
that developers with more then one version of emscripten often have
their cache cleared whenever they switch version.

With this method the cache it always local the to emscripten directory.
If distro what to install emscripten in a read-only or shared location
then can still use the CACHE config file setting, or set EM_CACHE
in the environment.

This is part of a wider push to stop using $HOME where we can and
move towards an embedded mode where it makes sense.

See: #9543