Build C11 thread API function in libpthread
This change doesn't do much other than compile in the relevant parts of
musl.

For all functions exception thrd_detach we simply sompile the musl code.
For thrd_detach musl normally aliases this to pthread_detech in the
pthread_detach.c but we implement pthread_detech so add the alias there.

Fixes #9242