Refactor and rename confusing pthreads async functions (#10772)
As noted in #9086, emscripten_async_queue_on_thread is a
misleading name since it is only async if calling another thread -
if we are already on the same thread, we just do the call
immediately.

Fixing that by making it actually async isn't an option, since
if the current thread never returns to the main event loop,
an async event would never happen. We have to fire it
synchronously in that case, but we don't know if that is the
case or not (it depends on user code). And it would be less
efficient for the common case.

As a solution emscripten_async_queue_on_thread has been renamed to
emscripten_dispatch_to_thread which no longer implies that it is async -
the operation is in fact only async if it is sent to another thread, while it
is sync if on the same one. A new emscripten_dispatch_to_thread_async
function is added which is always async.

That is a breaking change for people using that API. I would
guess not many have since while it is in threading.h it isn't
documented elsewhere. And this PR should help some of those
users (including the one that suggested this PR), so overall
it seems worth a breaking change.

Fixes #9086