Handle memory growth in node mmap (#11144)
This is the same as we have in MEMFS, but for NODEFS: when
we grow, the old buffer would be invalidated, so we have to check
for that. We couldn't find a nicer way there, and I can't think of a better
way now. Luckily this is a rare pattern (pass around a buffer reference
around a malloc).

Found by asan, but this is not an asan bug but due to how much
memory asan needs.