Fix the return value of __stdio_seek (#11145)
The ifdefs here were confusing, and in effect we had

#if emscripten
  if (..)
#else
  if (..)
    return -1;
#endif
  return 0;

So we ended up with

  if (..)
  return 0;

and no return otherwise. That can end up with the wrong value
when the seek fails.

Found by asan, but this is not an asan bug but due to how much
memory asan needs - for some reason the bug only occurs on
a failing seek with a large enough address.