Avoid printf for libcxxabi internal assertion message reporting (#11150)
Just trap for those messages. We ship libc++abi in our
builds on the web, while native doesn't, so their default
of verbose logging makes sense there, but for us it's
better to avoid printf and stdio just for this. As we have
proper stack traces on the web anyhow, and in a debug
build inlining won't confuse them, these should be pretty
understandable errors anyhow. (Also, these assertions
catch very rare cases, and perhaps are mostly useful
for libc++abi devs.)

Fixes #11142