Add useful errors on bad use of Promise instead of instance in new MODULARIZE (#11118)
It used to be possible to do

Module().onRuntimeInitialized = ...

but with new MODULARIZE (#10697) Module() returns a Promise
instead. The Promise of course allows a property to be placed on it, but
nothing would happen, so it's potentially annoying for users. With
this change, when ASSERTIONS are on we will throw an error
with an explanation, something like

You are setting onRuntimeInitialized on the Promise object, instead
of the instance. Use .then() to get called back with the instance, see
the MODULARIZE docs in src/settings.js

Likewise, it used to be possible to do

var instance = Module();
// .. later ..
instance.exportedThing()

but again, the "instance" is a Promise now, and the exports aren't
there. This will crash, but ".. is not a function" is not that helpful.
With this change it will give an explanation of what's wrong.