With EMTEST_SAVE_DIR clean the test output directory between runs
This change the default behavior.  Note that it doesn't remove the
directory itself, it only empties it out.  This means that if you have
a shell open to that directory, it won't disappear on you.

If you really want to keep dir content between runs you can still use
EMTEST_SAVE_DIR=2.