Only legalize as part of autodebug if we should do so (#11182)
Doing so unconditionally breaks standalone + autodebug.

This has to move some emcc.py code to decide if we need to legalize
to an earlier place, so we know that when we do the autodebug stuff
as part of the binaryen passes.

also_with_impure_standalone_wasm has to use BigInt support
while doing the "impure" part. In that mode we test standalone but
not in a wasm vm, rather in a JS VM. So we don't legalize, but we
do connect to JS. Without BigInt, we would trap on i64s.