Get cache lock before trying to empty it in Cache.clean()
The old implementation seemed to be resulting in some deadlock
issues.  With this method we play a little nicer by acquiring the
lock before cleaning the cache and we avoid re-creating the
lock file itself.

The deadlock was visible when running gen_struct_info right after
a new installation on windows (Noticed as part of #11192). This
change fixed the deadlock present in the testing of that PR.