Deprecate passing EM_CONFIG without a file
This is feature that I don't think has any users and removing
can simplify out code and testing matrix.

It was originally added back 2012 when it seems that per-project
configuration was a desirable feature:
99c46475c1944c1e62e04f30bd0db3233639515d

I think these days this should not be necessary, and even if it
if we still support `--em-config` on the command line.