Revert "Delete EH specification in test_dlmalloc_partial (#11201)"
This reverts commit 20529bb67559fea50a9389d9ba7548c66671f271.

The reason it started failing was actually a bug.
https://github.com/llvm/llvm-project/commit/d94bacbcf87a06abc0c1fc3405406399460debc3
started to warn on exception specifications with types, but it should
have done that only when wasm EH is used, i.e., `-fwasm-exceptions` is
enabled. This should not generate a warning when no EH is used or
Emscripten EH is used. This test does not use wasm EH, so it should've
not generate an error even if it has an exception specification.
https://github.com/llvm/llvm-project/commit/48acac36293805cef10b448d49d336da7db7bfa9
fixes this. Anyway, I don't think this particular test needs that
exception spec, but feel it's better to restore the old state, because
the error shouldn't have been there in the first place.