Add tests for preload-file and locateFile
For preload-file:
Test that putting .js, .wasm and .data files in a sub-directory
below the .html file still works. As stated in the reference,
emscripten should expect to find .data and .wasm files next to
the .js file.

For locateFile:
Test locateFile when using preload-file instead of manually
running the file packager.

See issue #9637