Make -msse explicitly require -msimd128
Passing -msse without -msimd128 doesn't make sense anyway, because the
SSE intrinsics glue will fail to compile when trying to call the WASM
SIMD intrinsics.

More importantly, this unbreaks projects that test for -msse in a
configure check, not expecting it to A) target an unstable version of
the WASM spec; and B) fail to actually compile the SSE code.