Fix and clarify optional license emitting (#11277)
Handle optional license emitting based on
EMIT_EMSCRIPTEN_LICENSE in a single central place
at the very end. This adds an extra processing step of
the final JS, but is unavoidable if we want to ensure that
no license is emitted by default even in -O0.

Also fixes emitting of multiple licences which could happen
due to bugs in the previous code, which didn't take into
account that closure would combine licenses with different
years, and that the license text varies slightly in different
places.