Fix clashing racing landing between #11329 and #11277 (#11331)
One PR added license handling, which ended up hitting a bad
case in the minimal runtime fixup code. We only remove the
license at the very end now, but to let the minimal runtime
fixups work as always, remove it earlier in that code path.

For fastcomp we also did not even run the final license handling
code, but now we need to or else we emit the license all the
time. This PR makes us run it, but that may regress some
corner cases of fastcomp source maps support (removing a
few lines of the license may confuse the parser). This removes
one test for source maps on fastcomp, but leaves the others.