wasm2c support: compile everything into a single C file (#11213)
The new WASM2C option makes us run wabt's wasm2c, combine the output
with an emscripten main() file + runtime in C, and emits a final
single C file of all that stuff. If that is built to a native executable it
does the same thing as the wasm(+js) would have done.

This has several potentially interesting use cases:

*   If we can build the emsdk tools (llvm, clang, binaryen, etc.) to C files,
    we can distribute those to users as an alternative for people that
    can't use the emsdk builds for some reason. For example if binaryen
    upgrades to c++20 but someone is on a linux distro that is too old for
    our standard builds, and doesn't have a c++20 compiler, then giving
    them a C file may help them.

*   A C build is also easier to generate. We have infrastructure for
    building on the 3 major OSes but it takes a bunch of work. A C build
    could be generated on a developer's machine very easily and run
    on all targets. This is the same as making a build that runs in node.js,
    which also would have those benefits. A C build has different
    tradeoffs, the main benefit being that it may be faster and in particular
    faster to start up (in theory node with wasm code caching could be
    similar, but that doesn't work atm - hopefully it will be possible eventually).

*   C builds are also interesting for sandboxing untrusted C code in
    a simple way.

Note that this PR uses an emscripten build of wasm2c, running on
node, instead of us needing to package normal binaries of wabt
(which is sort of an example of the second item in that list).