Fix wasm2c/os.c on MacOS and NetBSD (#11370)
Also add a test on the mac bot for wasm2c. (This will
conflict with #11364 which renames that test.)

Also simplify the disabling of wasm2c tests on windows - do it
in a single central place.

Re: windows, from reading online it sounds hard to get clang
working by itself on windows, and we have not done that on
our CI apparently as all our tests that would run clang to
build native executables are instead disabled there. So probably not
worth trying to do that just for this. Instead, maybe once wasm2c
works with msvc we can use that.