Add an assertion for bad --pre-js / files interaction (#11266)
--preload-file / --embed-file add some pre-js code to load the file. A user
pre-js can break that if it does Module = {} for example, as the files have
assigned to Module.preRun to set up a task.

This is rare because preloading files is really just common on the web, and
there we normally don't do Module = { .. } in the JS. Instead the Module
is defined on the HTML earlier.

However, I debugged something now that ended up being this, and it can
be confusing. To avoid others being confused, add some assertions that
check that Module.preRun has the right values even after the user pre-js.