Remove doublequote_commands option from get_building_env (#11391)
This argument was only used in a single place: in tests/runner.py
when constructing the environment to pass to `building.configure`
or `building.make`.

However the primary users of `building.configure` and `building.make`
that our users use are in `emconfigure` and `emmake` and here we
don't pass this option.

If this (windows-only) is actually needed then it should be used
in the production environment too, not just during testing.  Assuming
all the tests pass I think this is not needed.