Automatically resume AudioContexts on user input in SDL and OpenAL (#10843)
Tested on interactive.test_sdl_audio_beeps and
interactive.test_openal_buffers

Replaces #10700

We may also want to add an API as discussed there. But I
realized while starting to work on that that it's even better
to just fix those two API implementations. However, an API
might be useful for people creating their own audio systems,
so if there is demand we can do that too.