Embind + Asyncify
This adds a `val::await()` method to Embind API, which allows to await
dynamically retrieved `Promise`s from JavaScript.

This commit also adds an underlying `Asyncify.handlePromise` helper,
which is currently a thin wrapper around `Asyncify.handleSleep`
but might be extended over time, and useful anyway due to ubiquity
of `Promise`-based APIs on the Web.