Asyncify: Support for new add-list, and updates for binaryen list changes (#11438)
Add support for the new ASYNCIFY_ADD_LIST, and update existing list names
following the updates in Binaryen, so that now we have ASYNCIFY_ADD_LIST to
add a function, ASYNCIFY_REMOVE_LIST to remove one (previously this was
called ASYNCIFY_BLACKLIST), and ASYNCIFY_ONLY_LIST to set a list of the
only functions to instrument and no others (previously this was called
ASYNCIFY_WHITELIST).

The updated lists also handle indirect calls properly, so that if you
use ASYNCIFY_IGNORE_INDIRECT and then add (using either the
add-list or the only-list) all the functions that are on the stack when
pausing, then things will work (for more, see
WebAssembly/binaryen#2913). A new test is
added to show this.