Fix wasm0.test_emscripten_scan_registers (#11479)
That test assumed C locals were in wasm registers, but that is not
true in LLVM -O0 - they are on the stack. We didn't notice this because
they are also in wasm locals, temp locals that have short lifetimes.
With the new liveness analysis in binaryen, that was no longer
enough to make the test pass because we would ignore them.