wasm2js: Use normal constructors for Module, Instance, Memory (#11490)
Before we returned a new object. That means that instanceof does not
report the instances are proper instances of the functions, which I noticed
while working on wasm2js+pthreads. This fixes that and also shrinks code
size a little bit.

However, for Table I left it as it was before, since that has the .length
property which we'd need to use .defineProperty() or other fanciness,
which would also increase code size.