mmap() fixes: properly size and zero out (#11567)
Fixes #11565

The issue noticed there is that we don't expand to the page size, and don't
zero out memory. Those are fixed here. It's hard to test that though as those
details are really only noticeable internally or through undefined behavior
IIUC (e.g. we could mmap, munmap, then mmap again and check that it
was zeroed, but we'd need to ensure it wasn't just given a new block of
memory by malloc which is zeroes anyhow - other ideas?).

Also fixes other.test_mmap_and_munmap which this showed was
depending on not zero-initializing: it was off by one, and got lucky that
we didn't zero and also we got malloced the exact same region as before,
where the data we expected was present.