Wasm2js: Pthreads/Atomics support (#11505)
(Most of the hard work was on the binaryen side.)

This gets all the tests we run on asm.js pthreads to pass on wasm-backend +
wasm2js pthreads. That is just a subset of the full pthreads tests, so there may
still be issues, but as the main use case here is for debugging (it's easier to
debug a wasm2js build in general than a wasm one, and we want that for
pthreads too) this may be enough.

Another current limitation is that this disables full JS minification, because the
format wasm2js emits for the scratch helper functions and other runtime code
for pthreads is more complex than minifyGlobals can understand atm. We
could do some work to fix that, but again, if this is mostly for debugging then
it may not be needed.

A third limitation is that we can't use a .mem file, as the segments are
passive. That means we emit them in a flexible form inside the JS, where
we can support all the bulk memory operations. But in JS it is not as compact
as a binary .mem file.