Load ports dynamically based on directory contents (#11593)
To make this work I had to add explicit dependencies to each port so
that they can be built in the correct order.  Previously we were relying
on the ports being listed in a specific order to work around this issue.

This paves the way for ports to be maintained out of tree, and for
them to be added/removed by simply adding/removing files from this
directory.