Always warn/error about undefined symbols if they are explicitly exported (#11604)
ERROR_ON_UNDEFINED_SYMBOLS and WARN_ON_UNDEFINED_SYMBOLS exist
to allow build that contains undefined symbols to build created
with the missing symbols supplied at runtime somehow.

However, I think we should have a separate warning system for when function
that are explicitly exported by the user via EXPORTED_FUNCTIONS are missing
from a build.  In that case we always want to fail or warn.