Writes a test for new MODULARIZE error handling
The test asserts on the Promise being properly rejected with the
expected error. Reports the error message to the server to have a
clearer error message in failure cases, instead of the typical integer
reporting value used in other tests.

- Adds a helper assertion function for URL-decoding reported results.
- Makes browser test c++ fixture only report a result when the
report_result header file is included, to avoid duplicate result reports
to the test server.