Better roll prep (#11641)
The last commit was not good enough for LLVM to roll in. For some
reason other.test_emcc_v was disabled, which missed that it fails,
and it failed on chromium CI...

This skips it in a way that also skips on chromium CI, and that is
easier to undo later.