Do not consider embind contexts to be leaked
LeakSanitizer was previously reporting leaks of embind contexts, but
the reports were not actionable for users because the allocations were
inside the embind library. The context leaks are intentional because
they are allocated only once and live for the lifetime of the program,
so this patch uses the lsan public interface to tell lsan to ignore
the leaked contexts.

Testing this change also uncovered a bug that postRun hooks were
skipped when EXIT_RUNTIME was enabled, so this patch contains a
drive-by fix for that as well.

Fixes #11533.