Fix MAIN_THREAD_ASYNC_EM_ASM stack allocation (#11602)
EM_ASM are all variadic, so the params are on the C stack. The proxying
just naively proxies over arguments, including the parameter LLVM gave us
with the stack pointer - but that stack allocation may have been released and
the memory reused by the time the main thread gets the async message.

Instead, just unpack the varargs arguments immediately, which this PR
does by just moving the call to readAsmConstArgs to be first. This
this actually ends up simplifying the code - removing an ifdef on
the receiving side where we no longer need anything special.

Fixes #11350