Remove compiler-rt/filelist.txt (#11645)
This means that we will now build all compiler-rt builtins rather
than just a subset (at least with the llvm backend).

As part of this change I removed int128_builtins.cpp from the libc++
build.  This a pure duplicate of a compiler-rt function that was added
to libc++ for compatibility with gcc's runtime which is missing this
particular function.   Its only a problem for us if we force include
the entire library (e.g. when dynamic linking).