Increase slop in minimal runtime code size test for wasm2js (#11657)
Also as of July 16, 2020, wasm2js files have different sizes on
different platforms: Windows and MacOS improved to give a slightly
better thing than Linux does, which didn't change. This just
started to happen on CI, not in response to a code update, so it
may have been present all along but just noticed now. It only
happens in wasm2js, so it may be platform-nondeterminism in closure
compiler. It started on chromium CI for LLVM, but that doesn't
seem to be to blame as there is 0 difference in the wasm files,
just in the wasm2js of the same.

Anyhow, it's pretty small, and still better than the fastcomp slop.
Adding to the existing TODO there.