Fix padded double handling in EM_ASM varargs reading. (#11669)
Fixes #11665 which is a regression from #11227.

The main bug fixed here is that doubles must be aligned. The backend
will emit padding in varargs to ensure that, so we need to skip that
padding.

Also, it looks like previously the code pre-bumped buf for a double, instead of
bumping it after. That is, if the pointer was to address 8, then it would bump
it to 12, read from 12 >> 3 which is the same as 8 >> 3. Then the loop
normally bumps anyhow by 4 to 16. So we ended up reading from the right
place, and also the next thing after was read in the right place - but we
depended on rounding down of the >> operator. That is confusing, and
also it would trigger a false positive in SAFE_HEAP, so this PR avoids
that.