Add memset to emscripten_with_builtin_malloc
This was previously causing crashes due to ASan-instrumented memset
being used beneath the ASan boundary in the implementation of
mmap. Unfortunately, these crashes only happened when the allocation
sequence set the heap up in such a manner that the
previously-instrumented memset would intersect a poison region on the
heap, so we don't have a reproducer small enough to make a good test.

As a drive-by, also remove the use of __sys_pipe from libasan so that
it can link successfully with NO_FILESYSTEM.