Followup on addition of emscripten_stack_get_free
This is a followup on #11437 which added this API.  Sadly today the
linker doesn't know the final memory layout so until we fix #11773
this extra `init` function is needed.

The alternative, which is to modify binaryen and emscripten to inject
the stack end value post-link is IMHO way to much complexity to justify
the convenience of not having the call this init function.

Once we fix #11773 the init function can become uneeded/no-op.