Honor `-shared` flag
Today in emscripten "shared libraries" are just object files
and we assume we are linking a shared library wheneven the
file extension doesn't look like an exectuable.

Honoring this flag means that we can force this behavior
regardless of file extension.

This is the first in the chain of changes that seeks to make
emscripten less magical around linking shared libraries.

See #11795