JS optimizer: properly minify special JS names like "constructor" (#11790)
Just checking minifiedNames["constructor"] always returns true, since
it's built into each JS object. We need to use hasOwnProperty to do it
properly, as it want to ignore stuff up the prototype chain.

Also remove an if in function name minification - we should not check if
there is a minified name, as if we are minifying, they must all be minified,
or that's a bug.

Fixes #11740 which is a codebase that has a function called "constructor".