Store exceptions metadata in wasm memory (#11518)
Fixes #11503

This allocates memory in the heap for the relevant data structures,
and JS just refers to those places. This helps make exception handling
thread-safe (however, still not 100%, see comments in the code).

This also fixes issues with the old code mixing up catch contexts
with exceptions. The new code separates out the two properly. As
a result this fixes #11776 

Further improvements after this may include rewriting this to C
so it's entirely in wasm, which could make it 100% thread-safe and
also work in standalone mode etc. (However, the new wasm
exception handling will have those properties already.)