CI: Don't build all libraries and freeze the cache on Windows and Mac (#11841)
On those platforms we just run a subset of tests, which don't need most
of those libraries, so it's not helpful to build everything up front (which it
is on Linux, where we run the great majority of our tests in multiple jobs).

Before this window took 39 minutes to build and 5 to test, and with this
it takes 6 and 6 (so the test time increases very little, and most of the
build time is gone).

On mac the win is less extreme, it was 23 to build and 13 to test, and
with this it takes 2 to build and 20 to test (so test increases significantly,
but still worth it given the big build speedup).

A downside is that we don't test building libraries on those OSes. As
it's done mostly in pure python anyhow, this is probably not a big
deal (e.g. harfbuzz is an exception and uses cmake, but we hope to
change that too eventually).

Fixes #11840