SDL2_mixer: add support for WAV music (fix #11655)
Unintuitivelly, SDL2_mixer is being built without support for WAV music
~ through Mix_LoadMUS() ~, even though support for WAV samples ~ through
Mix_LoadWAV() ~ is always present.

One option for fixing this would be to add a "wav" option to
SDL2_MIXER_FORMATS, to enable WAV music support, but that could still
cause confusion, as, unlike other formats, it would only affect music,
not samples.

Since the size increase is about 8K (~10%) for the lib and 4K (~0.3%)
for the test program linked against it, enable support for WAV music
uncondittionally.