Fix bad interaction between MODULARIZE, ASan, O0, and Closure
MODULARIZE with a non-default export name combined with ASan, -O0, and
closure compiler had a bug where
__EMSCRIPTEN_PRIVATE_MODULE_EXPORT_NAME_SUBSTITUTION__ would fail to
be replaced properly with the module name. This was because ASan
causes the asanify JS transformation to be run, which changes the code
formatting and broke the regex that did the substitution. This PR
fixes the problem by making the regex robust to the presence of
whitespace.