Clarify what the failure case is in test_asan_modularized_with_closure (#11892)
I was surprised that .then() was needed to show the crash, since it should
be doing nothing... turns out that the bug was that Module() was returning
undefined, so it's a call that should do nothing, but it throws on calling a
method on something that is not an object.

This PR modifies the test to clarify what is being tested - check explicitly
that a proper Promise object is returned. I verified that this fails without
the fix in #11889 and passes with it.