Remove fastcomp-only ALIASING_FUNCTION_POINTERS setting (#11882)
Also removes the code block around it in emscripten.py. That's an obvious
easy win as it's a fastcomp-only function, and it's faster to remove such a
thing (when it's easily identifiable like here) rather than let future PRs remove
it line by line. Also removes the call chain to that function, also of fastcomp-
only functions (also easy wins, and necessary due to flake8).

See #11860