Remove obsolete and misleading wasm debugging docs. fixes #11919 (#11944)
This entire section was no longer useful. It was written under the assumption
that asm.js was stable and wasm was not, and how to debug stuff in wasm
given that. Today wasm is the most stable (and asm.js was replaced by wasm2js).

(And there is a separate Debugging page for debugging tips in general.)