Replace RESERVED_FUNCTION_POINTERS with ALLOW_TABLE_GROWTH (#11926)
Related to #11860 but this is NOT an NFC change. It has a very minor
noticeable effect.

The old RESERVED_FUNCTION_POINTERS let you add a fixed number
of slots for addFunction. We needed that in fastcomp where arbitrary
growth wasn't possible. As fastcomp is gone, we don't need the option
any more, and people can just use ALLOW_TABLE_GROWTH which
does the same but without a fixed limit (and there is no downside to
not having a limit).

Concretely this just maps RESERVED_FUNCTION_POINTERS to
ALLOW_TABLE_GROWTH. This has slightly different semantics, in
that it allows unlimited growth now. But that should be a strict
improvement, preventing previous error cases. Still, I added a mention
to the changelog.

The improvement to the code size tests is because of the simplification
in src/wasm2js.js. It basically makes release builds not contain extra
error logging which apparently we had earlier, and became obvious
while removing the old option's support.